+++ src/shuriken/multi.clj
@@ -1,5 +1,13 @@
 (ns shuriken.multi)
 
-(defn call-multi [multi dispatch-val & args]
+(defn call-method [multi dispatch-val & args]
   (apply (-> multi methods (get dispatch-val))
          args))
+
+(defn super-method [multi dispatch-val & args]
+  (let [ps (or (parents dispatch-val) #{:default})
+        _ (assert (not (> (count ps) 1))
+                  (format "Multiple parents found for dispatch val %s: %s"
+                          dispatch-val ps))
+        p (first ps)]
+    (apply call-method multi p args)))